Create calc.py
This version of the Python calculator used tkinter which allows the results of each calculation to be copied to the clipboard. The original program used pyperclip to copy the files to the clipboard but required the user to install it separably. This version bypasses that requirement and allows the user to run the file with only Python 3 installed.